+++ CHANGELOG.md
@@ -9,6 +9,10 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+### Added
+
+- Added `preemptible` argument to `TaskSpec.new()`.
+
 ## [v1.26.11](https://github.com/allenai/beaker-py/releases/tag/v1.26.11) - 2024-05-10
 
 ### Added